+++ jsdoc/index.js
@@ -1,6 +1,10 @@
 var path = require('canonical-path');
 var Package = require('dgeni').Package;
 
+/**
+ * @dgPackage jsdoc
+ * @description Tag parsing and extracting for JSDoc-based documentation
+ */
 module.exports = new Package('jsdoc', [require('../base')])
 
 // Add in extra pseudo marker processors
@@ -23,12 +27,13 @@ module.exports = new Package('jsdoc', [require('../base')])
 .factory(require('./services/transforms/whole-tag'))
 .factory(require('./services/transforms/trim-whitespace'))
 
+.factory(require('./services/jsParser'))
 .factory(require('./file-readers/jsdoc'))
 
 // Configure the processors
 
 .config(function(readFilesProcessor, jsdocFileReader) {
-  readFilesProcessor.fileReaders = [jsdocFileReader];
+  readFilesProcessor.fileReaders = [jsdocFileReader].concat(readFilesProcessor.fileReaders || []);
 })
 
 .config(function(parseTagsProcessor, getInjectables) {